requirements.txt
# Core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0

# HTTP & async
requests>=2.31.0
httpx>=0.27.0

# MongoDB (certifi for Atlas TLS handshake)
pymongo>=4.6.0
certifi>=2024.0.0

# Browser automation (optional; not available on Render free tier)
playwright>=1.40.0

# API server
fastapi>=0.109.0
python-multipart>=0.0.6
uvicorn[standard]>=0.27.0
jinja2>=3.1.0

# Scheduling
apscheduler>=3.10.0

# Timezone
pytz>=2024.1

# Optional for analytics
pandas>=2.0.0
